object: Sink floating refs on construction
Floating references don't really make sense outside of code written in
C, so we need to convert them into a normal ref on object construction.
The "transfer floating" annotation is just an alias for "transfer none"
so we detect such instances by checking if we're dealing with a
constructor and if the return value's transfer type is none.